Major modifications to code, changed sprites, changed enemy move logic, and started skeleton for enemy laser shots as well as player lives. Next up, finish enemy lasers and add in walls.

+++ Enemy.cs
@@ -11,7 +11,7 @@ namespace SpaceInvaders
         {
             public Texture2D Texture;
             public Vector2 Position;
-            public float Speed = 1;
+            public double Speed = 1;
             public bool Active;
 
             public int Width => Texture.Width;
